requirements.txt
python-dotenv
selenium
cassiopeia